CHANGE: Cleaned up the code for ComparisonHC
Threw out unused code (different linkage methods and Oracles).
Later, this might be desired back in (and can be added without a problem
through the constructor of ComparisonHC), but for now I will stick with the
LinkageMethod that works (the others achieved very bad results even on
very simple cluster sets).